paystream/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
boto3==1.34.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.0
websockets==12.0
